+++ .container/cli/cxcli/__init__.py
@@ -1,0 +1,2 @@
+"""CXSun cloud helper CLI package."""
+